requirements.txt
python-dotenv>=1.0.1
yt-dlp>=2025.4.30
ffmpeg-python>=0.2.0
pyacoustid>=1.3.0
pandas>=2.2.2
pytest>=8.2.0
